✨feat: store sensor data

+++ Application/Mogeun/app/src/main/java/io/ssafy/mogeun/data/Emg.kt
@@ -12,6 +12,6 @@ data class Emg (
     val deviceId: Int,
     @ColumnInfo(name = "sensing_part")
     val sensingPart: String,
-    val value: Double,
+    val value: Int,
     val time: Long
 )
